+++ profile/js/profile.user.js
@@ -1,12 +1,12 @@
-app.validUser().catch(() => {
-  app.redirectWithPreloader('../register')
-})
-
 const inputData = [{
   type: 'select',
   default: null,
   min: 10,
   clkId: 'theme',
+  saveNeeded: false,
+  preloaderNeeeded: false,
+  valueMatchRule: null,
+  isTextarea: false,
   data: [{
     title: 'Darkmoon',
     value: 'darkmoon',
@@ -15,15 +15,17 @@ const inputData = [{
     }
   }, {
     title: 'Red House',
-    value: 'red',
+    value: 'redhouse',
+    attr: {}
+  }, {
+    title: 'Blue Ocean',
+    value: 'blueocean',
+    attr: {}
+  }, {
+    title: 'Classic',
+    value: 'classic',
     attr: {}
   }]
-}, {
-  type: 'text',
-  default: '',
-  min: 10,
-  clkId: 'username',
-  data: []
 }]
 
 function InpDataToOptData(inpData) {
@@ -41,58 +43,121 @@ function InpDataToOptData(inpData) {
 }
 
 function useInputSelectors(inpData = inputData) {
+  function handleOutput(val, item) {
+    
+    switch (item.clkId) {
+      case 'theme':
+        if (val) {
+          themeManager.setTheme(val + '-theme')
+          themeManager.storeCurrentTheme()
+        }
+        break;
+    }
+  }
+
   inpData.forEach((item) => {
     var itemElem = document.getElementById(item.clkId);
     if (itemElem) {
       itemElem.onclick = function() {
         var title = itemElem.querySelector('.opt-title').innerText;
         if (item.type === 'select') {
-          modal.optionsPicker(InpDataToOptData(item), true, userboxUI.create('#(title)', '', '', userboxUI.input('#(inputId)', true)), title)
+          modal.optionsPicker(InpDataToOptData(item), true, userboxUI.pickerBox('#(inputId)', '#(title)', '', true), title).then(function(val) {
+            handleOutput(val, item)
+          })
         } else {
-          modal.prompt(title, item.default, title, false, item.type);
+          modal.prompt(title, item.default, title, false, item.type).then(function(val) {
+            handleOutput(val, item)
+          })
         }
       }
     }
   })
 }
 
-useInputSelectors()
+app.validUser().then((user) => {
+  inputData.push({
+    type: 'text',
+    default: user.fullname,
+    min: 0,
+    clkId: 'username',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  }, {
+    type: 'tel',
+    default: user.phone,
+    min: 8,
+    clkId: 'phone',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  }, {
+    type: 'password',
+    default: CryptoJS.AES.decrypt(user.password, config.ENC_KEY).toString(CryptoJS.enc.Utf8),
+    min: 8,
+    clkId: 'password',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  }, {
+    type: 'date',
+    default: user.dob,
+    min: 0,
+    clkId: 'dob',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  }, {
+    type: 'text',
+    default: user.address,
+    min: 10,
+    clkId: 'address',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: true
+  }, {
+    type: 'text',
+    default: user.goal,
+    min: 5,
+    clkId: 'goal',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: true
+  }, {
+    type: 'number',
+    default: user.height,
+    min: 30,
+    clkId: 'height',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  }, {
+    type: 'number',
+    default: user.weight,
+    min: 20,
+    clkId: 'weight',
+    data: [],
+    saveNeeded: true,
+    preloaderNeeeded: true,
+    valueMatchRule: null,
+    isTextarea: false
+  });
 
-function optPickerHtmlCreate(){
-  return new TagString(
-    `<div class="user-box user-select">
-  <div class="user-row">
-    <div class="user-icon user-left-input">
-      <input type="checkbox" name="" id="" value="" />
-    </div>
-    <div class="user-body">
-      <div class="user-name">#(title)</div>
-      <div class="user-subtext">subtext</div>
-    </div>
-  </div>
-</div>`
-    )
-  
-  return new TagString(
-      `
-      <div class="option-item-box">
-        <div class="option-item-left">
-          <input type="checkbox" class="option-item-input">
-        </div>
-        <div class="option-item-center">
-          <div class="option-item-title">Hello</div>
-          <div class="option-item-text">Demo</div>
-        </div>
-      </div>
-      `
-    )
-}
-
-modal.optionsPicker([
-  {
-    inputId: 'tya',
-    value: 'wj',
-    attr: {},
-    title: 'hrllo',
-    text: 'js'
-  }], false, optPickerHtmlCreate())
+  useInputSelectors()
+}).catch(() => {
+  app.redirectWithPreloader('../register')
+})